NOTE - I had to rename main.cpp to game_main.cpp because otherwise you have two compilation units named "main".  This should have no visible impact.
These changes are intended to minimize the impact on the other platform builds.  I have faked unistd.h and raceintospace_config.h in the visual studio project.  (Not added here)

FIXES - Overflowing initialzing the VAB structure in BuildVAB.

CHANGES - Structure alignment is now handled with #prgama push( pack, 1 )/#pragma pop instead of __alignment__ for cross-compiler compatibility.

+++ src/game/mis_m.cpp
@@ -45,7 +45,7 @@ extern struct mStr Mis;
 
 extern uint16_t MisStat;
 extern char pCnt, tMen, BIG;          // Counter for pAry
-extern int16_t pAry[15];         /**< Array for Presige Firsts compelted */
+extern int pAry[15];         /**< Array for Presige Firsts compelted */
 extern REPLAY Rep;
 
 void Tick(char);
